requirements.txt
langchain
python.dotenv
ipykernel
langchain-community
langchain-groq
streamlit